Polish: reject null inlineFiles, Math.toIntExact, ArrayList accumulator
* Message.user(content, inlineFiles) rejects null inlineFiles outright
  rather than coercing to List.of() — callers pass an explicit empty
  list when there are none. Records don't tolerate null components by
  spec §22.3.

* StopClassifier converts state.currentTurnIndex() to int via
  Math.toIntExact instead of a silent (int) cast. maxTurns is int, so
  the cast is practically safe today, but Math.toIntExact asserts the
  invariant.

* SessionOptions.Builder.withHook used to allocate an ArrayList + run
  List.copyOf per call (quadratic when building hooks one at a time).
  Now appends to a long-lived ArrayList accumulator; the canonical
  constructor calls List.copyOf at build() time anyway.

+++ session/src/main/java/ai/singlr/session/loop/StopClassifier.java
@@ -95,7 +95,7 @@ public final class StopClassifier {
       return Optional.of(
           new ResultMessage.ErrorMaxTurns(
               state.sessionId(),
-              (int) state.currentTurnIndex(),
+              Math.toIntExact(state.currentTurnIndex()),
               state.usage(),
               state.cost(),
               state.elapsed()));
